[fix]offsetZ to fpx value
